requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0

# LangChain 生态（按官方拆包方式）
langchain-core>=0.3.70,<0.4
langchain-community>=0.3.25,<0.4
langchain-openai>=0.3.25,<0.4
langchain-text-splitters>=0.3.8,<0.4
langchain-chroma>=0.2.4,<0.3
langchain-huggingface>=0.1.2,<1.0

# 向量库与嵌入模型
chromadb>=1.0.15,<2.0
sentence-transformers>=3.0,<6.0

# 文档解析
pypdf>=5.0,<7.0
docx2txt>=0.8,<1.0
python-docx>=1.1,<2.0

# 其他
python-multipart>=0.0.9,<1.0
pydantic>=2.7,<3.0
markdown>=3.4,<4.0
playwright>=1.40,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
